pyproject.toml
[tool.poetry]
name = "daily-allocation-tasks-automation"
version = "0.1.0"
description = ""
authors = ["Falak097 <falakgala09@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.165.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
apscheduler = "^3.11.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
